MK-220 fix: Show quick reply from other user
다른 사람이 보낸 [이상 무 보고]의 응답 버튼이 보이지 않는 버그 수정
